Refactor the files and classes
