+++ src/classroom.rb
@@ -1,0 +1,12 @@
+class Classroom
+  attr_accessor :label, :students
+
+  def initialize(label)
+    @label = label
+    @students = []
+  end
+
+  def add_student(student)
+    @students << student 
+  end
+end
