+++ src/decorator.rb
@@ -1,0 +1,20 @@
+require_relative './nameable'
+
+class Decorator < Nameable
+  attr_reader :component
+
+  def initialize(nameable)
+    super()
+    @nameable = nameable
+  end
+
+  def correct_name
+    @nameable.correct_name
+  end
+end
+
+class TrimmerDecorator < Decorator
+  def correct_name
+    @nameable.correct_name[0, 10]
+  end
+end
