+++ src/factory.rb
@@ -29,6 +29,8 @@ class Factory < UserPrompt
   private
 
   def show_allowed_options
+    puts "please select on of the following options by typing the number:\n\n"
+    
     options = [
       'Create person',
       'Create classroom',
