+++ src/main.rb
@@ -9,7 +9,7 @@ def main
     when 1..8
       factory.process(choice)
     when 9
-      puts 'App quitting gracefully..'
+      puts 'App is closing...'
       break
     else
       puts 'INVALID OPTION'
