+++ src/teacher.rb
@@ -1,8 +1,8 @@
 require './person'
 
 class Teacher < Person
-  def initialize(age, name, specialization, parent_permission: true)
-    super(age, name, parent_permission: parent_permission)
+  def initialize(age, name, specialization)
+    super(age, name)
     @specialization = specialization
   end
