+++ src/user_prompt.rb
